Moved DTOs to separate folder
