+++ api/Entities/DTO/TimeDTO.cs
@@ -1,0 +1,21 @@
+namespace api.Entities
+{
+    public class TimeDTO : Time
+    {
+        public TimeDTO(Time time)
+        {
+            if (time != null)
+            {
+                TimeHour = time.TimeHour.ToString(@"hh\:mm");
+                Remarks = time.Remarks;
+            }
+            else
+            {
+                TimeHour = null;
+            }
+
+        }
+
+        public new string? TimeHour { get; set; }
+    }
+}
