+++ api/Entities/Time.cs
@@ -12,23 +12,3 @@ namespace api.Entities
         public string? Remarks { get; set; }
     }
 }
-
-namespace api.Entities
-{
-    public class TimeDTO : Time
-    {
-        public TimeDTO(Time time)
-        {
-            if (time != null)
-            {
-                TimeHour = time.TimeHour.ToString(@"hh\:mm\:ss");
-            }
-            else
-            {
-                TimeHour = null;
-            }
-        }
-
-        public new string? TimeHour { get; set; }
-    }
-}
